[minor] Rename rotations.h to slerp.h
