+++ src/CMakeLists.txt
@@ -46,7 +46,7 @@ set(CONTROLLER_HDR
     ${CMAKE_CURRENT_SOURCE_DIR}/../include/lipm_walking/utils/LowPassVelocityFilter.h
     ${CMAKE_CURRENT_SOURCE_DIR}/../include/lipm_walking/utils/clamp.h
     ${CMAKE_CURRENT_SOURCE_DIR}/../include/lipm_walking/utils/polynomials.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/../include/lipm_walking/utils/rotations.h)
+    ${CMAKE_CURRENT_SOURCE_DIR}/../include/lipm_walking/utils/slerp.h)
 
 set(CONTROLLER_NAME "${PROJET_NAME}_controller")
